requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
python-dotenv>=1.0
pyarrow>=14.0
tqdm>=4.66
pytest>=7.4
